requirements.txt
discord.py-self
Unidecode
networkx
